pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Second project with Hexlet - gendifferencies"
authors = ["Oleg Kharitonov <o_kharitonov@mail.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^5.0.0"
pyyaml = "^6.0.1"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
